+++ CHANGELOG.md
@@ -1,3 +1,25 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] 
 
-* TODO: Describe initial release.
+* Initial.
+
+## [0.0.2] 
+
+* Added Example
+
+## [0.0.3] 
+
+* Updated Example and Documents.
+
+
+## [0.0.4] 
+
+* Minor fixes.
+
+## [0.0.5] 
+
+* Updated Example and Documents.
+
+
+
+
+